tests/requirements_tests.txt
fastapi[all]
httpx
pillow
boto3==1.28.17
botocore==1.31.17
torch
torchvision
prometheus_client
s3fs==2023.6.0
fsspec==2023.6.0
pytest
pillow==10.2.0
